+++ components/AgentMesssage.tsx
@@ -1,0 +1,3 @@
+export default function AgentMessage({ agentMessage }: { agentMessage: string }) {
+	return <p class="text-brain-text text-lg pb-3">{ agentMessage }</p>;
+}