+++ components/UserMessage.tsx
@@ -1,0 +1,3 @@
+export default function UserMessage({userMessage}: {userMessage: string}) {
+	return <p class="text-bubble-gum-text text-lg">{ userMessage }</p>;
+}