+++ routes/_app.tsx
@@ -8,7 +8,7 @@ export default define.page(function App({ Component }) {
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>FreshDemo</title>
       </head>
-      <body>
+      <body f-client-nav>
         <Component />
 			</body>
     </html>