+++ routes/about.tsx
@@ -1,12 +1,13 @@
+import { Partial } from "fresh/runtime";
 import { Countdown } from "../islands/Countdown.tsx";
 import { define } from "../utils.ts";
 
 export default define.page(() => {
 	return (
-		<main>
+		<Partial name="mylink">
 			<h1>About</h1>
 			<p>This is the about page.</p>
 			<Countdown />
-		</main>
+		</Partial>
 	);
 });