+++ routes/conversation.tsx
@@ -1,0 +1,22 @@
+import { Partial } from "fresh/runtime";
+import { define } from "../utils.ts";
+import Conversation from "../components/Conversation.tsx";
+import ChatForm from "../components/ChatForm.tsx";
+
+export default define.page((ctx) => {
+	const userMessage = ctx.url.searchParams.get("userMessage") as string;
+	console.log(typeof userMessage)
+	const agentMessage = "Let's watch this partial go through!";
+	const formKey = crypto.randomUUID();
+
+	return (
+		<div>
+			<Partial name="chatResponse" mode="append">
+				<div class="max-w-[50dvw]">
+					<Conversation userMessage={userMessage} agentMessage={agentMessage}/>
+				</div>
+			</Partial>
+			<Partial name="chatResponseForm"><ChatForm url="/conversation" key={formKey}/></Partial>
+		</div>
+	)
+})