+++ CHANGELOG.md
@@ -1,4 +1,11 @@
+# Versions
+
 - **v-0.0.1**
     - Initial Commit.
     - Added CHANGELOG.md
     - Made some styling rules.
+
+-**v-0.1.0**
+    - Added ChatForm.tsx to components.
+    - Added Conversation.tsx to components.
+    - Added conversation.tsx to routes which renders the Conversation JSX element in a partial and Chatform JSX element in a Partial so that the form resets on every submission despite being inside a partial.